+++ examples/bid-optimizer/src/index.ts
@@ -1,8 +1,7 @@
 import { core } from "@mediarithmics/plugins-nodejs-sdk";
-import { MyActivityAnalyzerPlugin } from './MyPluginImpl'
+import { MyBidOptimizerPlugin } from './MyPluginImpl'
 
 // All the magic is here
-const plugin = new MyActivityAnalyzerPlugin();
+const plugin = new MyBidOptimizerPlugin();
 const runner = new core.ProductionPluginRunner(plugin);
-
 runner.start();