+++ src/core/class/mediarithmics/activity-analyzer/ActivityAnalyzerBasePlugin.ts
@@ -67,10 +67,10 @@ export abstract class ActivityAnalyzerPlugin extends BasePlugin {
     const activityAnalyzer = results[0];
     const activityAnalyzerProps = results[1];
 
-    const context = {
+    const context:ActivityAnalyzerBaseInstanceContext = {
       activityAnalyzer: activityAnalyzer,
       activityAnalyzerProperties: activityAnalyzerProps
-    } as ActivityAnalyzerBaseInstanceContext;
+    };
 
     return context;
   }