+++ src/core/interfaces/mediarithmics/api/bid-optimizer/BidOptimizerInterface.ts
@@ -1,0 +1,14 @@
+export interface BidOptimizerResponse {
+    status: string;
+    data: BidOptimizer;
+    count: number;
+}
+
+export interface BidOptimizer {
+    id: string;
+    organisation_id: string;
+    name: string;
+    engine_artifact_id: string;
+    engine_version_id: string;
+    engine_group_id: string;
+}