+++ src/core/interfaces/mediarithmics/api/common/ValueInterface.ts
@@ -1,5 +1,5 @@
 export interface Value {
-  value?: string | number;
+  value?: string | number | boolean;
   id?: string;
   version?: string;
   url?: string;
@@ -7,4 +7,6 @@ export interface Value {
   original_file_name?: string;
   asset_id?: string;
   file_path?: string;
+  last_modified?: number;
+  recommender_id?: string;
 }