+++ src/tests/AdRendererBaseTests.ts
@@ -125,7 +125,6 @@ describe("Ad Contents API test", function() {
           count: 45,
           data: [
             {
-              id: "42",
               technical_name: "hello_world",
               value: {
                 value: "Yay"